refactor(workflows): merge lock-db.ts into utils.ts
Move isFolderEffectivelyLockedDb and isWorkflowEffectivelyLockedDb into
utils.ts which already imports @sim/db and contains other server-side
workflow helpers. Eliminates the non-standard -db suffix file naming
and consolidates duplicate imports across 5 route files.
