+++ apps/sim/app/api/workflows/[id]/variables/route.ts
@@ -7,8 +7,10 @@ import { z } from 'zod'
 import { AuditAction, AuditResourceType, recordAudit } from '@/lib/audit/log'
 import { checkSessionOrInternalAuth } from '@/lib/auth/hybrid'
 import { generateRequestId } from '@/lib/core/utils/request'
-import { isWorkflowEffectivelyLockedDb } from '@/lib/workflows/lock-db'
-import { authorizeWorkflowByWorkspacePermission } from '@/lib/workflows/utils'
+import {
+  authorizeWorkflowByWorkspacePermission,
+  isWorkflowEffectivelyLockedDb,
+} from '@/lib/workflows/utils'
 import type { Variable } from '@/stores/variables/types'
 
 const logger = createLogger('WorkflowVariablesAPI')
